bugfix for downModis._getToday()
BUG: if self.today or self.enday are not strings, i.e. they are
	datetime.date objects, then they could be passed to
	str2date() which would cause an error in that function.
	This could happen if, for example, self.getListDays() is
	called more than once.

FIX: changed 'if None: ... else: str2date()' statements to
	'if None: ... elif type() == str: str2date()'

A similar, and perhaps better way to fix this bug would be to
run the contents of _getToday() during __init__ and remove the
refernce to _getToday() inside getListDays(). However, this could
become a problem if self.today or self.enday are modified after
instantiation.

	modified: pymodis/downmodis.py